+++ src/lib/IdCardText/index.ts
@@ -1,0 +1,11 @@
+import DidPlc from "./DidPlc.svelte";
+import Nip05 from "./Nip05.svelte";
+import OpenpgpFpr from "./OpenpgpFpr.svelte";
+import Plain from "./Plain.svelte";
+
+export default {
+  Plain,
+  Nip05,
+  DidPlc,
+  OpenpgpFpr,
+};
