--- src/lib/LinkCardText/index.ts
@@ -1,11 +1,0 @@
-import LinkCardTextDidPlc from "./LinkCardTextDidPlc.svelte";
-import LinkCardTextNip05 from "./LinkCardTextNip05.svelte";
-import LinkCardTextOpenpgpFpr from "./LinkCardTextOpenpgpFpr.svelte";
-import LinkCardTextPlain from "./LinkCardTextPlain.svelte";
-
-export default {
-  Plain: LinkCardTextPlain,
-  Nip05: LinkCardTextNip05,
-  DidPlc: LinkCardTextDidPlc,
-  OpenpgpFpr: LinkCardTextOpenpgpFpr,
-};
